feat(Airtable): Add parcel as a bundler
