+++ package.json
@@ -44,6 +44,8 @@
     "lint": "lerna exec -- yarn lint"
   },
   "devDependencies": {
+    "@parcel/packager-ts": "^2.0.0",
+    "@parcel/transformer-typescript-types": "^2.0.0",
     "lerna": "^4.0.0",
     "prettier": "^2.3.2",
     "typescript": "^4.4.2"
