+++ packages/airtable/package.json
@@ -2,8 +2,10 @@
   "name": "@bothrs/airtable",
   "repository": "git@github.com:bothrs/open-source.git",
   "version": "0.2.4",
+  "source": "src/index.ts",
   "main": "dist/index.js",
-  "module": "dist/index.js",
+  "module": "dist/module.js",
+  "types": "dist/types.d.ts",
   "publishConfig": {
     "access": "public"
   },
@@ -28,8 +30,9 @@
     "dist"
   ],
   "scripts": {
-    "build": "yarn build:clean && tsc",
-    "build:clean": "rimraf \"./dist\"",
+    "watch": "parcel watch",
+    "build:clean": "rimraf \"dist\"",
+    "build": "yarn build:clean && parcel build",
     "test": "jest --config=\"../../jest.config.js\" --roots=\"./packages/airtable/src\" --coverage --runInBand --passWithNoTests"
   },
   "peerDependencies": {
@@ -41,6 +44,7 @@
     "jest": "^27.2.4",
     "jest-extended": "^0.11.5",
     "node-fetch": "^3.0.0",
+    "parcel": "^2.0.0",
     "query-string": "^7.0.1",
     "rimraf": "^3.0.2",
     "ts-jest": "^27.0.5",
